Made two versions of SelectGeodesics
SelectGeodesics_one_side.m only returns the vector with the number of
geodesics to plot for each max.
SelectGeodesics_two_side.m returns the vector with the number of
geodesics to plot AS WELL AS adding a fourth column in maxconnect which
says how many endpoints the geodesic was selected from (0, 1 or 2).